requirements.txt
setuptools==65.0.0
wheel~=0.37.1
python-daemon
